requirements.txt
numpy
pandas
scikit-learn
torch
faiss-cpu
pyarrow
tqdm